zmq_device/requirements.txt
pyzmq==26.0.3